Aggregate#2 - Location aggregate design added
